+++ sqlparser/test/parser/create_trigger_test.dart
@@ -12,7 +12,7 @@ final _block = Block([
   ]),
   SelectStatement(
     columns: [StarResultColumn()],
-    from: [TableReference('tbl')],
+    from: TableReference('tbl'),
   ),
 ]);