+++ sqlparser/test/parser/select/generic_test.dart
@@ -32,18 +32,16 @@ final Map<String, AstNode> testCases = {
         expression: SubQuery(
           select: SelectStatement(
             columns: [StarResultColumn(null)],
-            from: [TableReference('table2', null)],
+            from: TableReference('table2', null),
           ),
         ),
       ),
     ],
-    from: [
-      TableReference('tbl', null),
-    ],
+    from: TableReference('tbl', null),
   ),
   'SELECT * FROM tbl WHERE id IN ()': SelectStatement(
     columns: [StarResultColumn(null)],
-    from: [TableReference('tbl', null)],
+    from: TableReference('tbl', null),
     where: InExpression(
       left: Reference(columnName: 'id'),
       inside: Tuple(expressions: []),
@@ -53,7 +51,7 @@ final Map<String, AstNode> testCases = {
     columns: [
       ExpressionResultColumn(expression: Reference(columnName: 'rowid')),
     ],
-    from: [TableReference('tbl')],
+    from: TableReference('tbl'),
   ),
 };